requirements.txt
flash 
gunicorn
pandas
google-generativeai
python-dotenv